Renamed JavaSitter file name to java_sitter.py to be PEP8 compliant
